+++ tensorflow/core/common_runtime/mkl_layout_pass.cc
@@ -46,6 +46,7 @@ limitations under the License.
 #include "tensorflow/core/lib/hash/hash.h"
 #include "tensorflow/core/platform/logging.h"
 #include "tensorflow/core/util/mkl_heuristics.h"
+#include "tensorflow/core/util/onednn_env_vars.h"
 #include "tensorflow/core/util/tensor_format.h"
 #include "tensorflow/core/util/util.h"
 
@@ -370,6 +371,7 @@ class MklLayoutRewritePass : public GraphOptimizationPass {
     csinfo_.tanh = "Tanh";
     csinfo_.tanh_grad = "TanhGrad";
     csinfo_.reshape = "Reshape";
+    csinfo_.sparse_matrix_matmul = "SparseMatrixMatMul";
     csinfo_.slice = "Slice";
     csinfo_.softmax = "Softmax";
     csinfo_.split = "Split";
@@ -540,6 +542,12 @@ class MklLayoutRewritePass : public GraphOptimizationPass {
     rinfo_.push_back({csinfo_.matmul,
                       mkl_op_registry::GetMklOpName(csinfo_.matmul),
                       CopyAttrsAll, MatMulRewrite, kRewriteForOpNameChange});
+#ifdef ENABLE_ONEDNN_V3
+    rinfo_.push_back(
+        {csinfo_.sparse_matrix_matmul,
+         mkl_op_registry::GetMklOpName(csinfo_.sparse_matrix_matmul),
+         CopyAttrsAll, SparseMatrixMatMulRewrite, kRewriteForOpNameChange});
+#endif
     rinfo_.push_back({csinfo_.leakyrelu,
                       mkl_op_registry::GetMklOpName(csinfo_.leakyrelu),
                       CopyAttrsAll, LeakyReluRewrite, GetRewriteCause()});
@@ -985,6 +993,7 @@ class MklLayoutRewritePass : public GraphOptimizationPass {
     string tanh_grad;
     string transpose;
     string reshape;
+    string sparse_matrix_matmul;
     string slice;
     string softmax;
     string split;
@@ -1541,6 +1550,50 @@ class MklLayoutRewritePass : public GraphOptimizationPass {
     }
     return false;
   }
+
+  static bool SparseMatrixMatMulRewrite(const Node* n) {
+    DataType T;
+    const TensorProto* proto = nullptr;
+    Tensor tensor;
+    bool adjoint_a, adjoint_b, transpose_a, transpose_b, transpose_out;
+
+    // Check the environment variable
+    if (!UseOnednnSpmm()) {
+      VLOG(2) << "TF_ENABLE_ONEDNN_SPMM is disabled";
+      return false;
+    } else {
+      VLOG(2) << "TF_ENABLE_ONEDNN_SPMM is enabled";
+    }
+
+    // Check the datatype
+    TF_CHECK_OK(GetNodeAttr(n->def(), "T", &T));
+    if (T != DT_FLOAT) {
+      VLOG(2) << "_MklSparseMatrixMatMul only supports DT_FLOAT";
+      return false;
+    }
+
+    // Check for adjointing
+    TF_CHECK_OK(GetNodeAttr(n->def(), "adjoint_a", &adjoint_a));
+    TF_CHECK_OK(GetNodeAttr(n->def(), "adjoint_b", &adjoint_b));
+    if (adjoint_a || adjoint_b) {
+      VLOG(2)
+          << "_MklNativeSparseMatrixMatMul doesn't support adjointing matrices";
+      return false;
+    }
+
+    // Check transposing
+    TF_CHECK_OK(GetNodeAttr(n->def(), "transpose_a", &transpose_a));
+    TF_CHECK_OK(GetNodeAttr(n->def(), "transpose_b", &transpose_b));
+    TF_CHECK_OK(GetNodeAttr(n->def(), "transpose_output", &transpose_out));
+    if (transpose_a || transpose_b || transpose_out) {
+      VLOG(2) << "_MklNativeSparseMatrixMatMul doesn't support transposing "
+                 "matrices";
+      return false;
+    }
+
+    return true;
+  }
+
   // For oneDNN, only int32 is supported for axis data type
   static bool ConcatV2Rewrite(const Node* n) {
     DataType T;
