+++ tensorflow/core/graph/mkl_graph_util.h
@@ -143,6 +143,7 @@ inline string GetMklNativeOpName(const string& name) {
   // prefixed with _Mkl instead of _MklNative.
   bool result =
       (0 == name.compare("ConjugateTranspose") ||
+       0 == name.compare("SparseTensorDenseMatMul") ||
        0 == name.compare("BatchMatMul") || 0 == name.compare("BatchMatMulV2") ||
        0 == name.compare("Einsum") || 0 == name.compare("MatMul") ||
        0 == name.compare("Transpose") || 0 == name.compare("QuantizeV2") ||
