+++ tensorflow/core/kernels/mkl/mkl_matmul_ops_common.h
@@ -21,7 +21,6 @@ limitations under the License.
 #include <string>
 #include <vector>
 
-#include "unsupported/Eigen/CXX11/Tensor"  // from @eigen_archive
 #include "dnnl.hpp"
 #include "tensorflow/core/framework/op.h"
 #include "tensorflow/core/framework/op_kernel.h"
@@ -29,6 +28,7 @@ limitations under the License.
 #include "tensorflow/core/kernels/mkl/mkl_kernel_util.h"
 #include "tensorflow/core/util/mkl_util.h"
 #include "tensorflow/core/util/onednn_env_vars.h"
+#include "unsupported/Eigen/CXX11/Tensor"  // from @eigen_archive
 #if defined(DNNL_AARCH64_USE_ACL) && defined(ENABLE_ONEDNN_OPENMP)
 #include "tensorflow/core/platform/mutex.h"
 #endif
@@ -776,6 +776,7 @@ struct MklMatMulParams {
   memory::dims a_strides;
   memory::dims b_strides;
   memory::dims c_strides;
+  memory::dim a_nnz;
   struct PostOpParam {
     string name;
     std::vector<float> param;
@@ -787,17 +788,19 @@ struct MklMatMulParams {
 
   MklMatMulParams(string prefix, memory::dims a_dims, memory::dims b_dims,
                   memory::dims c_dims, memory::dims a_strides,
-                  memory::dims b_strides, memory::dims c_strides)
+                  memory::dims b_strides, memory::dims c_strides,
+                  memory::dim a_nnz = 0)
       : prefix(prefix),
         a_dims(a_dims),
         b_dims(b_dims),
         c_dims(c_dims),
         a_strides(a_strides),
         b_strides(b_strides),
-        c_strides(c_strides) {}
+        c_strides(c_strides),
+        a_nnz(a_nnz) {}
 };
 
-template <typename Tlhs, typename Trhs, typename Toutput>
+template <typename Tlhs, typename Trhs, typename Toutput, bool CSR = false>
 class MklMatMulPrimitive : public MklPrimitive {
  public:
   explicit MklMatMulPrimitive(const MklMatMulParams& params)
@@ -811,9 +814,12 @@ class MklMatMulPrimitive : public MklPrimitive {
   dnnl::memory::desc GetScratchPadDesc() {
     return context_.prim_desc->scratchpad_desc();
   }
+
   void Execute(const std::shared_ptr<stream>& stream, const Tlhs* a_data,
                const Trhs* b_data, const Toutput* c_data, void* sp_data,
-               void* mul_data = nullptr, void* add_data = nullptr) {
+               void* mul_data = nullptr, void* add_data = nullptr,
+               const int32_t* a_col_indices = nullptr,
+               const int32_t* a_row_pointers = nullptr) {
 #if defined(DNNL_AARCH64_USE_ACL) && defined(ENABLE_ONEDNN_OPENMP)
     mutex_lock lock(primitive_execution_mu_);
 #endif
@@ -824,20 +830,29 @@ class MklMatMulPrimitive : public MklPrimitive {
         static_cast<void*>(const_cast<Trhs*>(b_data)), *stream);
     context_.c_mem->set_data_handle(
         static_cast<void*>(const_cast<Toutput*>(c_data)), *stream);
-    context_.sp_mem->set_data_handle(sp_data, *stream);
 
+    if (sp_data != nullptr) context_.sp_mem->set_data_handle(sp_data, *stream);
     if (mul_data != nullptr)
       context_.mul_mem->set_data_handle(mul_data, *stream);
     if (add_data != nullptr)
       context_.add_mem->set_data_handle(add_data, *stream);
 #else
-    context_.a_mem->set_data_handle(
-        static_cast<void*>(const_cast<Tlhs*>(a_data)));
+    if (CSR) {
+      context_.a_mem->set_data_handle(
+          static_cast<void*>(const_cast<Tlhs*>(a_data)), 0);
+      context_.a_mem->set_data_handle(
+          static_cast<void*>(const_cast<int32_t*>(a_col_indices)), 1);
+      context_.a_mem->set_data_handle(
+          static_cast<void*>(const_cast<int32_t*>(a_row_pointers)), 2);
+    } else {
+      context_.a_mem->set_data_handle(
+          static_cast<void*>(const_cast<Tlhs*>(a_data)));
+    }
     context_.b_mem->set_data_handle(
         static_cast<void*>(const_cast<Trhs*>(b_data)));
     context_.c_mem->set_data_handle(
         static_cast<void*>(const_cast<Toutput*>(c_data)));
-    context_.sp_mem->set_data_handle(sp_data);
+    if (sp_data != nullptr) context_.sp_mem->set_data_handle(sp_data);
     if (mul_data != nullptr) context_.mul_mem->set_data_handle(mul_data);
     if (add_data != nullptr) context_.add_mem->set_data_handle(add_data);
 #endif  // !ENABLE_ONEDNN_OPENMP && !ENABLE_ONEDNN_V3
@@ -847,7 +862,7 @@ class MklMatMulPrimitive : public MklPrimitive {
     context_.a_mem->set_data_handle(DummyData);
     context_.b_mem->set_data_handle(DummyData);
     context_.c_mem->set_data_handle(DummyData);
-    context_.sp_mem->set_data_handle(DummyData);
+    if (sp_data != nullptr) context_.sp_mem->set_data_handle(DummyData);
     if (mul_data != nullptr) context_.mul_mem->set_data_handle(DummyData);
     if (add_data != nullptr) context_.add_mem->set_data_handle(DummyData);
   }
@@ -907,8 +922,16 @@ class MklMatMulPrimitive : public MklPrimitive {
     std::shared_ptr<dnnl::primitive> matmul_primitive = nullptr;
 
     // Create MatMul descriptor and primitive descriptor.
-    context_.a_md.reset(new memory::desc({params.a_dims}, MklDnnType<Tlhs>(),
-                                         params.a_strides));
+    if (CSR) {
+      // If it's a CSR matrix
+      const auto tmp = memory::desc::csr(
+          params.a_dims, MklDnnType<Tlhs>(), params.a_nnz,
+          dnnl::memory::data_type::s32, dnnl::memory::data_type::s32);
+      context_.a_md.reset(new memory::desc(tmp));
+    } else {
+      context_.a_md.reset(new memory::desc({params.a_dims}, MklDnnType<Tlhs>(),
+                                           params.a_strides));
+    }
 
     context_.b_md.reset(new memory::desc({params.b_dims}, MklDnnType<Trhs>(),
 #ifdef DNNL_AARCH64_USE_ACL
@@ -967,8 +990,13 @@ class MklMatMulPrimitive : public MklPrimitive {
 #endif  // !ENABLE_ONEDNN_V3
 
     // Create memory primitive based on dummy data.
-    context_.a_mem.reset(
-        new dnnl::memory(*context_.a_md, cpu_engine_, DummyData));
+    if (CSR) {
+      context_.a_mem.reset(new dnnl::memory(*context_.a_md, cpu_engine_,
+                                            std::vector<void*>(3, DummyData)));
+    } else {
+      context_.a_mem.reset(
+          new dnnl::memory(*context_.a_md, cpu_engine_, DummyData));
+    }
 #ifdef DNNL_AARCH64_USE_ACL
     context_.b_mem.reset(new dnnl::memory(
         context_.prim_desc.get()->weights_desc(), cpu_engine_, DummyData));
@@ -1019,24 +1047,25 @@ class MklMatMulPrimitive : public MklPrimitive {
 #endif
 };
 
-template <typename T, typename Tlhs, typename Trhs, typename Toutput>
+template <typename T, typename Tlhs, typename Trhs, typename Toutput,
+          bool CSR = false>
 class MklMatMulPrimitiveFactory : public MklPrimitiveFactory<T> {
  public:
-  static MklMatMulPrimitive<Tlhs, Trhs, Toutput>* Get(
+  static MklMatMulPrimitive<Tlhs, Trhs, Toutput, CSR>* Get(
       const MklMatMulParams& params, bool do_not_cache) {
-    MklMatMulPrimitive<Tlhs, Trhs, Toutput>* matmul_prim = nullptr;
+    MklMatMulPrimitive<Tlhs, Trhs, Toutput, CSR>* matmul_prim = nullptr;
 
     if (do_not_cache) {
       // Always create new primitive
-      matmul_prim = new MklMatMulPrimitive<Tlhs, Trhs, Toutput>(params);
+      matmul_prim = new MklMatMulPrimitive<Tlhs, Trhs, Toutput, CSR>(params);
     } else {
       // Try to find a suitable one in pool
-      matmul_prim = dynamic_cast<MklMatMulPrimitive<Tlhs, Trhs, Toutput>*>(
-          MklMatMulPrimitiveFactory<T, Tlhs, Trhs, Toutput>::GetInstance()
+      matmul_prim = dynamic_cast<MklMatMulPrimitive<Tlhs, Trhs, Toutput, CSR>*>(
+          MklMatMulPrimitiveFactory<T, Tlhs, Trhs, Toutput, CSR>::GetInstance()
               .GetMklMatMul(params));
       if (matmul_prim == nullptr) {
-        matmul_prim = new MklMatMulPrimitive<Tlhs, Trhs, Toutput>(params);
-        MklMatMulPrimitiveFactory<T, Tlhs, Trhs, Toutput>::GetInstance()
+        matmul_prim = new MklMatMulPrimitive<Tlhs, Trhs, Toutput, CSR>(params);
+        MklMatMulPrimitiveFactory<T, Tlhs, Trhs, Toutput, CSR>::GetInstance()
             .SetMklMatMul(params, matmul_prim);
       }
     }
