+++ tensorflow/core/util/onednn_env_vars.h
@@ -27,6 +27,8 @@ bool UseSystemAlloc();
 
 bool ThreadPoolUseCallerThread();
 
+bool UseOnednnSpmm();
+
 std::string FPMathModeSetting();
 }  // namespace tensorflow
 #endif  // INTEL_MKL
