+++ tensorflow/python/kernel_tests/linalg/sparse/csr_sparse_matrix_dense_mat_mul_onednn_grad_test.py
@@ -4,7 +4,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#   http://www.apache.org/licenses/LICENSE-2.0
+#     http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,7 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""CSR sparse matrix tests."""
+"""Tests for TF_ENABLE_ONEDNN_SPMM=1."""
+
+import os
 
 from tensorflow.python.platform import test
 from tensorflow.python.kernel_tests.linalg.sparse import csr_sparse_matrix_dense_mat_mul_grad_test_base
@@ -22,6 +24,7 @@ CSRSparseMatrixDenseMatMulGradTest = \
   CSRSparseMatrixDenseMatMulGradTest
 
 if __name__ == '__main__':
+  os.environ['TF_ENABLE_ONEDNN_SPMM'] = '1'
   csr_sparse_matrix_dense_mat_mul_grad_test_base.\
     define_csr_sparse_matrix_dense_mat_mul_grad_tests(
       CSRSparseMatrixDenseMatMulGradTest)
